make scratchpad identification more unique
add random string of letters to title
add class="Alacritty" to filter out
non-terminal cases.

The problem was: firefox always vanished
into a scratchpad whenever i searched for
"math" or "alsamixer" as then the title
of the window contained those keywords
and the scratchpad was identified
just by "math" and "alsamixer"

I hope that this is enough to practically
prevent such issues.